Move iteration calculations into methods of the Status class.
This means that the iterations can be invoked after initialization by
calling run() on the Status object, rather than passing the Status into
the Tsne's run() method. Users can then be sure that changes to the
Tsne's parameters do not affect the iterations of any existing Status
object returned by initialize().

This protection is mostly to avoid weird situations where the parameters
change halfway through the iterations. However, it is most important in
situations where the number of threads increases from 1 halfway through
an iteration, in which case the old approach would attempt to store
values in an empty parallel_buffer - resulting in a segfault.

As a consequence of this change, the set of iteration-relevant parameters
have been bundled up into a separate class and cloned into the Status
object during initialize(). This should also clarify (for us) which
parameters are needed in iteration vs initialization.

While I was here, I removed all the logging messages as well.