Switch to free functions, replace interpolation with leaf approximation. (#9)
Pretty much as it says on the tin; we replace the Tsne class with a free
initialize() function that accepts Options and returns a Status object. We
remove the top-level run() methods as it is simple to call Status.run(). We
also updated the initialize() overloads to use the new knncolle interfaces.

The SPTree underwent some major refactoring to streamline some of its logic:

- Fixed a bug that slightly overstated the contribution of the repulsive force
  from the leaf node containing the point of interest.
- Explicitly check for exact duplicates to avoid unnecessary recursion.
- Use the full width instead of the half-width in the BH approximation criteria.
  This aligns with external definitions of the BH algorithm, though the change
  has no actual effect here as we increase the default theta to 1 to compensate.

More tests have been added to double-check all of these new scenarios. 

The interpolation code has been replaced by a more natural "leaf approximation"
approach, whereby the SPTree's leaf nodes are used to estimate the respulsive
force based on their centers of mass. The respulsive force for each point is
then approximated by the force on its leaf node's center of mass plus the
repulsion between the point itself and the center of mass. This is more natural
as it re-uses the tree structure and supports approximation at any ndim_.

We converted the R bindings into a dedicated package for easier testing.
Also replaced the C++ gallery with a basic example in R for nicer plotting.