Only compute forces for leaf nodes during the leaf approximation.
Previously, we were computing forces for all nodes, which is unnecessary work
as the non-leaf node don't get used in the approximation. For parallelization,
we also make some effort to balance the work across threads, otherwise the last
thread will end up processing most of the leaves at the "end" of the tree.