Improve convergence of the Newton-based search for beta.
We improve convergence by setting the initial beta based on the distance to the
furthest neighbor. This adjusts the initial guess to match the scale of the
distances, reducing the number of iterations to get to the right ballpark.

We also make sure to update the binary search bounds, even when Newton's is
working properly. This ensures that, if we ever need to fall back to a binary
search, we can leverage previous Newton successes to narrow the interval.