Hardened sanity checks to work across more dimensions.
There is no theoretical guarantee that the simulated separation manifests in
the first dimension of the embedding (though in practice it always seems to do
so). So we just check all other dimensions just to be safe.

We also tighten up the sanity check with the leaf approximation, to avoid
having a hard-coded absolute difference. Instead, we just check that the
means of each group do not lie in the range of the other group.